Separate split_criteria from Training logic
Have a separate module where we implement different split_feature
critera, and have the train function on the TreeNode except any function
that has the correct signature.
