+++ src/tree_node.rs
@@ -1,4 +1,8 @@
 use crate::dataset::Dataset;
+use crate::split_criteria::gini_coeficient_split_feature;
+use crate::split_criteria::mean_squared_error_split_feature;
+use crate::split_criteria::SplitFunction;
+use crate::vec_utils;
 use std::cmp::Ordering::Equal;
 use std::collections::HashMap;
 use std::{cell::RefCell, rc::Rc};
@@ -13,129 +17,26 @@ pub struct TreeNode {
     right: Option<Rc<RefCell<TreeNode>>>,
 }
 
-#[derive(Debug, PartialEq)]
-struct SplitResult {
-    col_index: usize,
-    row_index: usize,
-    split: f32,
-    prediction: f32,
-    loss: f32,
-}
-
 impl TreeNode {
-    pub fn float_avg(x: &[f32]) -> f32 {
-        x.iter().sum::<f32>() / x.len() as f32
-    }
-
-    pub fn sort_two_vectors(a: &[f32], b: &[f32]) -> (Vec<f32>, Vec<f32>) {
-        let mut pairs: Vec<(&f32, &f32)> = a.iter().zip(b).collect();
-        pairs.sort_by(|&a, &b| a.0.partial_cmp(b.0).unwrap_or(Equal));
-
-        pairs.into_iter().map(|(x, y)| (*x, *y)).unzip()
-    }
-
-    fn split_feature_reg(col_index: usize, feature: &[f32], target: &[f32]) -> SplitResult {
-        let (sorted_feature, sorted_target) = TreeNode::sort_two_vectors(feature, target);
-
-        let mut row_index = 1;
-        let mut min_mse = f32::MAX;
-        let mut last = sorted_feature[0];
-
-        let square: Vec<f32> = sorted_target
-            .iter()
-            .map(|x| x * x)
-            .scan(0.0, |state, x| {
-                *state += x;
-                Some(*state)
-            })
-            .collect();
-        let sum: Vec<f32> = sorted_target
-            .iter()
-            .scan(0.0, |state, x| {
-                *state += x;
-                Some(*state)
-            })
-            .collect();
-
-        for i in 1..sorted_feature.len() {
-            if sorted_feature[i] > last {
-                //    var = \sum_i^n (y_i - y_bar) ** 2
-                //           = (\sum_i^n y_i ** 2) - n_samples * y_bar ** 2
-                //
-                let left_square_sum = square[i - 1];
-                let right_square_sum = square[square.len() - 1] - square[i - 1];
-
-                let left_avg = sum[i - 1] / (i as f32);
-                let right_avg = (sum[sum.len() - 1] - sum[i - 1]) / (sum.len() - i) as f32;
-
-                let right_mse = right_square_sum - (sum.len() - i) as f32 * right_avg * right_avg;
-                let left_mse = left_square_sum - i as f32 * left_avg * left_avg;
-                let mse = left_mse + right_mse;
-
-                if mse < min_mse {
-                    row_index = i;
-                    min_mse = mse;
-                }
-
-                last = sorted_feature[i];
-            }
-        }
-
-        SplitResult {
-            col_index,
-            row_index,
-            split: (sorted_feature[row_index] + sorted_feature[row_index - 1]) / 2.0,
-            prediction: TreeNode::float_avg(target),
-            loss: min_mse,
-        }
-    }
-
-    fn split_feature_clf(col_index: usize, feature: &[f32], target: &[f32]) -> SplitResult {
-        let (sorted_feature, sorted_target) = TreeNode::sort_two_vectors(feature, target);
-
-        let mut row_index = 1;
-        let mut min_gini = f32::MAX;
-        let mut last = sorted_feature[0];
-
-        let cumsum: Vec<f32> = sorted_target
-            .iter()
-            .scan(0.0, |state, x| {
-                *state += x;
-                Some(*state)
-            })
-            .collect();
-
-        for i in 1..sorted_feature.len() {
-            if sorted_feature[i] > last {
-                let left_cumsum = cumsum[i - 1] / (i as f32);
-                let right_cumsum =
-                    (cumsum[cumsum.len() - 1] - cumsum[i - 1]) / (cumsum.len() - i) as f32;
-
-                let gini = left_cumsum * (1.0 - left_cumsum) + right_cumsum * (1.0 - right_cumsum);
-
-                if gini < min_gini {
-                    row_index = i;
-                    min_gini = gini;
-                }
-
-                last = sorted_feature[i];
-            }
-        }
-
-        SplitResult {
-            col_index,
-            row_index,
-            split: (sorted_feature[row_index] + sorted_feature[row_index - 1]) / 2.0,
-            prediction: TreeNode::float_avg(target),
-            loss: min_gini,
-        }
-    }
-
     pub fn train(train: Dataset, curr_depth: i32, max_depth: i32) -> TreeNode {
+        Self::_train(
+            train,
+            curr_depth,
+            max_depth,
+            mean_squared_error_split_feature,
+        )
+    }
+
+    fn _train(
+        train: Dataset,
+        curr_depth: i32,
+        max_depth: i32,
+        split_feature: SplitFunction,
+    ) -> TreeNode {
         if (curr_depth == max_depth) | (train.target_vector.len() == 1) {
             return TreeNode {
                 split: None,
-                prediction: TreeNode::float_avg(&train.target_vector),
+                prediction: vec_utils::float_avg(&train.target_vector),
                 samples: train.target_vector.len(),
                 feature_name: None,
                 left: None,
@@ -146,10 +47,9 @@ impl TreeNode {
         let best_feature = train
             .feature_matrix
             .iter()
-            //.par_iter()
             .enumerate()
             .map(|(index, feature_vector)| {
-                TreeNode::split_feature_reg(index, feature_vector, &train.target_vector)
+                split_feature(index, feature_vector, &train.target_vector)
             })
             .min_by(|a, b| a.loss.partial_cmp(&b.loss).unwrap_or(Equal))
             .unwrap();
@@ -158,7 +58,7 @@ impl TreeNode {
         let mut right_dataset = train.clone_without_data();
 
         for i in 0..train.feature_names.len() {
-            let (_, sorted_feature) = TreeNode::sort_two_vectors(
+            let (_, sorted_feature) = vec_utils::sort_two_vectors(
                 &train.feature_matrix[best_feature.col_index],
                 &train.feature_matrix[i],
             );
@@ -170,7 +70,7 @@ impl TreeNode {
             right_dataset.feature_matrix.push(second_half);
         }
 
-        let (_, sorted_target) = TreeNode::sort_two_vectors(
+        let (_, sorted_target) = vec_utils::sort_two_vectors(
             &train.feature_matrix[best_feature.col_index],
             &train.target_vector,
         );
@@ -186,85 +86,23 @@ impl TreeNode {
             prediction: best_feature.prediction,
             samples: train.target_vector.len(),
             feature_name: Some(train.feature_names[best_feature.col_index].clone()),
-            left: Some(Rc::new(RefCell::new(TreeNode::train(
+            left: Some(Rc::new(RefCell::new(TreeNode::_train(
                 left_dataset,
                 curr_depth + 1,
                 max_depth,
+                split_feature,
             )))),
-            right: Some(Rc::new(RefCell::new(TreeNode::train(
+            right: Some(Rc::new(RefCell::new(TreeNode::_train(
                 right_dataset,
                 curr_depth + 1,
                 max_depth,
+                split_feature,
             )))),
         }
     }
 
     pub fn train_clf(train: Dataset, curr_depth: i32, max_depth: i32) -> TreeNode {
-        if (curr_depth == max_depth) | (train.target_vector.len() == 1) {
-            return TreeNode {
-                split: None,
-                prediction: TreeNode::float_avg(&train.target_vector),
-                samples: train.target_vector.len(),
-                feature_name: None,
-                left: None,
-                right: None,
-            };
-        }
-
-        let best_feature = train
-            .feature_matrix
-            .iter()
-            //.par_iter()
-            .enumerate()
-            .map(|(index, feature_vector)| {
-                TreeNode::split_feature_clf(index, feature_vector, &train.target_vector)
-            })
-            .min_by(|a, b| a.loss.partial_cmp(&b.loss).unwrap_or(Equal))
-            .unwrap();
-
-        let mut left_dataset = train.clone_without_data();
-        let mut right_dataset = train.clone_without_data();
-
-        for i in 0..train.feature_names.len() {
-            let (_, sorted_feature) = TreeNode::sort_two_vectors(
-                &train.feature_matrix[best_feature.col_index],
-                &train.feature_matrix[i],
-            );
-
-            let mut first_half = sorted_feature.clone();
-            let second_half = first_half.split_off(best_feature.row_index);
-
-            left_dataset.feature_matrix.push(first_half);
-            right_dataset.feature_matrix.push(second_half);
-        }
-
-        let (_, sorted_target) = TreeNode::sort_two_vectors(
-            &train.feature_matrix[best_feature.col_index],
-            &train.target_vector,
-        );
-
-        let mut first_half = sorted_target;
-        let second_half = first_half.split_off(best_feature.row_index);
-
-        left_dataset.target_vector = first_half;
-        right_dataset.target_vector = second_half;
-
-        TreeNode {
-            split: Some(best_feature.split),
-            prediction: best_feature.prediction,
-            samples: train.target_vector.len(),
-            feature_name: Some(train.feature_names[best_feature.col_index].clone()),
-            left: Some(Rc::new(RefCell::new(TreeNode::train(
-                left_dataset,
-                curr_depth + 1,
-                max_depth,
-            )))),
-            right: Some(Rc::new(RefCell::new(TreeNode::train(
-                right_dataset,
-                curr_depth + 1,
-                max_depth,
-            )))),
-        }
+        Self::_train(train, curr_depth, max_depth, gini_coeficient_split_feature)
     }
 
     pub fn predict_row(&self, row: &HashMap<&String, f32>) -> f32 {
@@ -306,7 +144,7 @@ impl TreeNode {
             .map(|(xt, xp)| (xt - xp).powf(2.0))
             .sum();
 
-        let avg = TreeNode::float_avg(x_true);
+        let avg = vec_utils::float_avg(x_true);
         let var: f32 = x_true.iter().map(|x| (x - avg).powf(2.0)).sum();
 
         1.0 - mse / var
@@ -325,26 +163,4 @@ impl TreeNode {
 #[cfg(test)]
 mod test {
     use super::*;
-
-    #[test]
-    fn test_sort_two_vectors() {
-        assert_eq!(
-            TreeNode::sort_two_vectors(&vec![2.0, 0.0, 1.0], &vec![-1.0, 0.0, 1.0]),
-            (vec![0.0, 1.0, 2.0], vec![0.0, 1.0, -1.0])
-        );
-    }
-
-    #[test]
-    fn test_split_feature() {
-        assert_eq!(
-            TreeNode::split_feature_reg(1, &vec![2.0, 0.0, 1.0], &vec![-1.0, 0.0, 1.0]),
-            SplitResult {
-                col_index: 1,
-                row_index: 2,
-                split: 1.5,  // takes the average between the value to split on and the previous
-                prediction: 0.0,
-                loss: 0.5,
-            }
-        );
-    }
 }
